pyproject.toml
[tool.poetry]
name = "college_football_elo"
version = "0.1.0"
description = ""
authors = ["Alexander Talbott <atalbott13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.1"
streamlit = "^1.22.0"
requests = "^2.30.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
